requirements.txt
### Usage ###
# LaTeX (and other) templates
jinja2
# numpy is optional, install if you want.
# numpy
#
### Development ###
pylint
#
## testing ##
pytest-mock
hypothesis
coverage
pytest-cov
#
## manually create documentation ##
sphinx
sphinxcontrib-napoleon
